# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.InvoiceLink:
    properties:
      account_name:
        type: string
      amount:
        type: integer
      created_at:
        $ref: '#/definitions/models.Timestamp'
      description:
        type: string
      document_id:
        type: integer
      document_type:
        description: bill, invoice, payout or recurring_payment_occurrence
        type: string
      id:
        type: integer
      reference:
        type: string
      transaction_date:
        type: string
      transaction_id:
        type: integer
    type: object
...
  handlers.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  models.Invoice:
    properties:
      allocated:
        type: integer
      amount:
        type: integer
      contact_id:
        type: integer
      contact_name:
        description: Computed fields
        type: string
      created_at:
        $ref: '#/definitions/models.Timestamp'
      due_date:
        type: string
      file_url:
        type: string
      id:
        type: integer
      invoice_number:
        type: string
      issue_date:
        type: string
      items:
        items:
          $ref: '#/definitions/models.InvoiceItem'
        type: array
      notes:
        type: string
      status:
        type: string
      unallocated:
        type: integer
      updated_at:
        $ref: '#/definitions/models.Timestamp'
    type: object
  models.InvoiceInput:
    properties:
      amount:
        type: integer
      contact_id:
        type: integer
      due_date:
        type: string
      file_url:
        type: string
      invoice_number:
        type: string
      issue_date:
        type: string
      items:
        items:
          $ref: '#/definitions/models.InvoiceItemInput'
        type: array
      notes:
        type: string
      status:
        type: string
    type: object
  models.InvoiceItem:
    properties:
      amount:
        type: integer
      created_at:
        $ref: '#/definitions/models.Timestamp'
      description:
        type: string
      id:
        type: integer
      invoice_id:
        type: integer
      quantity:
        type: number
      unit:
        type: string
      unit_price:
        type: integer
      updated_at:
        $ref: '#/definitions/models.Timestamp'
    type: object
  models.InvoiceItemInput:
    properties:
      amount:
        type: integer
      description:
        type: string
      quantity:
        type: number
      unit:
        type: string
      unit_price:
        type: integer
    type: object
...
  models.Timestamp:
    properties:
      time.Time:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API for managing accounts, contacts, bills, invoices, and transactions.
  title: Portal API
  version: 1.0.0
paths:
...
  /invoices:
    get:
      description: Get a list of all receivable invoices, with current status and
        allocation info.
      parameters:
      - description: Filter by contact (customer)
        in: query
        name: contact_id
        type: integer
      - description: Search by invoice number, notes, or customer name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Invoice'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List invoices
      tags:
      - invoices
    post:
      consumes:
      - application/json
      description: Create a new receivable invoice.
      parameters:
      - description: Invoice contents
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/models.InvoiceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Invoice'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create invoice
      tags:
      - invoices
  /invoices/{id}:
    delete:
      description: Remove an invoice.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete invoice
      tags:
      - invoices
    get:
      description: Get details and allocation status of a specific invoice.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Invoice'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get invoice
      tags:
      - invoices
    put:
      consumes:
      - application/json
      description: Update details of an existing invoice.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated invoice contents
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/models.InvoiceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Invoice'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update invoice
      tags:
      - invoices
  /invoices/{id}/items:
    get:
      description: Get all line items for a specific invoice.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.InvoiceItem'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: List invoice items
      tags:
      - invoices
    post:
      consumes:
      - application/json
      description: Add a new line item to an existing invoice.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      - description: Line item contents
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.InvoiceItemInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InvoiceItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create invoice item
      tags:
      - invoices
  /invoices/{id}/items/{itemId}:
    delete:
      description: Remove a line item from an invoice.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete invoice item
      tags:
      - invoices
    put:
      consumes:
      - application/json
      description: Update an existing line item in an invoice.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: Updated line item contents
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.InvoiceItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InvoiceItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update invoice item
      tags:
      - invoices
  /invoices/{id}/links:
    get:
      description: Get all payment transactions linked to a specific invoice.
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.InvoiceLink'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get invoice links
      tags:
      - invoices
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"